# Allow downstream projects to use faster regexp