--- a/model/labels/regexp.go
+++ b/model/labels/regexp.go
@@ -14,9 +14,10 @@
 package labels
 
 import (
-	"regexp"
-	"regexp/syntax"
 	"strings"
+
+	"github.com/grafana/regexp"
+	"github.com/grafana/regexp/syntax"
 )
 
 type FastRegexMatcher struct {