NOTE: This PR does not now change the regexp implementation used by Prometheus, just the package import path.

This PR makes it possible for downstream projects that use Prometheus code to take advantage of some optimisations I have submitted to the Go project:
* [regexp: allow patterns with no alternates to be one-pass](https://go-review.googlesource.com/c/go/+/353711)
* [regexp: speed up onepass prefix check](https://go-review.googlesource.com/c/go/+/354909)
* [regexp: handle prefix string with fold-case](https://go-review.googlesource.com/c/go/+/358756)
* [regexp: avoid copying each instruction executed](https://go-review.googlesource.com/c/go/+/355789)
* [regexp: allow prefix string anchored at beginning](https://go-review.googlesource.com/c/go/+/377294)

I have a [FOSDEM talk](https://fosdem.org/2022/schedule/event/go_finite_automata/) on Saturday 6th Feb about this. https://www.youtube.com/watch?v=iBTwVyfH9eY

I collected the optimisations into a new repo https://github.com/grafana/regexp; it exists purely to make PRs like this possible and can be retired once the changes land upstream.

I added a benchmark, largely using the patterns in existing tests - I didn't spend much time looking for patterns that flatter the optimisations:
```
name                                                old time/op  new time/op  delta
FastRegexMatcher/foo-4                               397ns ± 4%   187ns ± 2%  -52.86%  (p=0.008 n=5+5)
FastRegexMatcher/^foo-4                             37.7ns ± 2%  38.5ns ±10%     ~     (p=1.000 n=5+5)
FastRegexMatcher/(foo|bar)-4                         409ns ± 2%   366ns ± 2%  -10.47%  (p=0.008 n=5+5)
FastRegexMatcher/foo.*-4                            1.72µs ± 4%  1.61µs ± 2%   -6.30%  (p=0.008 n=5+5)
FastRegexMatcher/.*foo-4                            1.48µs ± 7%  1.40µs ± 4%     ~     (p=0.095 n=5+5)
FastRegexMatcher/^.*foo$-4                          1.52µs ± 5%  1.43µs ± 1%   -6.26%  (p=0.008 n=5+5)
FastRegexMatcher/^.+foo$-4                          1.51µs ± 7%  1.44µs ± 2%     ~     (p=0.056 n=5+5)
FastRegexMatcher/.*-4                               5.06µs ± 5%  4.86µs ± 3%     ~     (p=0.222 n=5+5)
FastRegexMatcher/.+-4                               5.18µs ± 6%  4.83µs ± 3%   -6.62%  (p=0.016 n=5+5)
FastRegexMatcher/foo.+-4                            1.72µs ± 4%  1.63µs ± 3%   -5.64%  (p=0.016 n=5+5)
FastRegexMatcher/.+foo-4                            1.46µs ± 4%  1.42µs ± 3%     ~     (p=0.222 n=5+5)
FastRegexMatcher/.*foo.*-4                          4.73µs ± 1%  4.59µs ± 2%   -3.08%  (p=0.008 n=5+5)
FastRegexMatcher/(?i:foo)-4                          386ns ± 3%   213ns ± 2%  -44.72%  (p=0.008 n=5+5)
FastRegexMatcher/(prometheus|api_prom)_api_v1_.+-4  55.8ns ± 3%  55.1ns ± 4%     ~     (p=0.548 n=5+5)
FastRegexMatcher/((fo(bar))|.+foo)-4                5.88µs ± 0%  5.57µs ± 3%   -5.27%  (p=0.016 n=4+5)
```

Note that some regexps are optimised away by `FastRegexMatcher` so don't see any benefit.