@roidelapluie 2022-02-02
Thanks. My concern here is the complexity of the issues at stake. I would rather let the go developers merge the patches and use the upstream version, currently it seems that upstream patches have not been peer-reviewed yet.