@bboreham 2022-02-03
Fair enough.  Go 1.18 is in beta; we are six months minimum from seeing the patches released upstream.

I missed one detail: my regexp repo has a `main` branch which is identical to Go upstream and a `speedup` branch which includes the patches, so it would be possible for Prometheus to merge this change but link `grafana/regexp` on main; downstream projects can `replace` with the speedup branch.