@bboreham 2022-02-03
> without strings.Cut

Confirmed. It is taken from the Go repo at commit 810b08b8ec, basically just before that change.