@roidelapluie 2022-02-04
Can you add regexp to depguard (in .golangci.yml) ?

Also, I'd like to clarify that we have the intent to switch back in ~1 year to the upstream regexp, once your patches have been released.