@bboreham 2022-03-04
I've updated `grafana/regexp` main branch for https://github.com/golang/go/issues/51112; will make a PR soon to update Prometheus.