@roidelapluie 2022-03-04
Changing the target would be enough. Thanks!

On 04 Mar 02:34, Bryan Boreham wrote:
> I can; shall I change the target of #10397 or make a second PR?
> 
> -- 
> Reply to this email directly or view it on GitHub:
> https://github.com/prometheus/prometheus/pull/10251#issuecomment-1059041095
> You are receiving this because you modified the open/close state.
> 
> Message ID: ***@***.***>

-- 
 (o-    Julien Pivotto
 //\    Open-Source Consultant
 V_/_   Inuits - https://www.inuits.eu